% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei.r
\name{add_spei_cols}
\alias{add_spei_cols}
\title{Add calculated PET (thornthwaite) and SPEI columns to a dataframe .}
\usage{
add_spei_cols(
  df,
  tcol,
  pcol,
  scale_mo = 12,
  na.rm = FALSE,
  plot = TRUE,
  loc = NULL
)
}
\arguments{
\item{df}{(dataframe) dataframe with a date index, temp, ppt and lat
(latitude) columns}

\item{tcol}{(string) column name for monthly temperature}

\item{pcol}{(string) column name for monthly precipitation total}

\item{scale_mo}{(int) time scale, or number of months used to calculate SPEI}

\item{na.rm}{(bool) flag to mask NA values from the calculation (default=F)}

\item{plot}{(bool) Flag for making diagnostic plot of SPEI}

\item{loc}{(string) Name of the site, which appears in the plot & varname}
}
\value{
df_out (dataframe) df with PET and SPEI columns added
}
\description{
Optionally make a plot of the data
}
\details{
Default frequency is monthly (freq=12) but there may be ways to make
weekly frequency work (freq=52) if weekly PET is available.
}
