% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei.r
\name{get_cwdiff}
\alias{get_cwdiff}
\title{Get climatic water difference from precip and pet data}
\usage{
get_cwdiff(precip, pet)
}
\arguments{
\item{precip}{(xts object) monthly precipitation sums}

\item{pet}{(xts object) monthly potential ET sums}
}
\value{
cwdiff (xts object) climatic water differential
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
