% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei.r
\name{get_spei}
\alias{get_spei}
\title{Get SPEI values from climatic water differential at given frequency and
integration period.}
\usage{
get_spei(cwdiff, freq = 12, int_per = 6, plot = TRUE)
}
\arguments{
\item{cwdiff}{(xts obj) monthly climatic water differential time series}

\item{freq}{(int) frequency of cwdiff estimates (12=monthly, 52=weekly)}

\item{int_per}{(int) number of periods (weeks, months) used to calculate SPEI}

\item{plot}{(bool) Flag for making diagnostic plot of SPEI}
}
\value{
spei_int (xts obj) spei time series at given integration period

spei_int_interp (xts obj) spei_int with NA values interpolated over
}
\description{
Optionally make a plot of the data
}
\details{
Default frequency is monthly (freq=12) but there may be ways to make
weekly frequency work (freq=52) if weekly PET is available.
}
