% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghcn.r
\name{ghcn_df_subset}
\alias{ghcn_df_subset}
\title{Subset a GHCN dataframe by station}
\usage{
ghcn_df_subset(ghcn_df, stationId)
}
\arguments{
\item{ghcn_df}{(tibble) GHCN dataframe with (usually) multiple stations}

\item{stationId}{(char) GHCN station identifier to keep}
}
\value{
ghcn_df2 (tibble) table of subsetted station data
}
\description{
This function drops all rows in a GHCN dataframe except for
the given station.
}
