% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghcn.r
\name{ghcn_inventory}
\alias{ghcn_inventory}
\title{Get the station identifiers list for GHCN}
\usage{
ghcn_inventory(
  loc = "https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt"
)
}
\arguments{
\item{loc}{(char) location of the file at NCEI, or elsewhere}
}
\value{
ghcn_inv (tibble) table of query results
}
\description{
This function fetches a table of GHCN weather station identifiers and
other information and returns data as a tibble.
}
\details{
Current location is:
https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt

API documentation:
https://www.ncei.noaa.gov/support/access-data-service-api-user-documentation
}
