% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghcn.r
\name{ghcn_summary_data}
\alias{ghcn_summary_data}
\title{Get a table of summary data for a GHCN station}
\usage{
ghcn_summary_data(
  stationIds,
  startdt,
  enddt,
  varnames = "TMAX,TMIN,TAVG,PRCP",
  summarytype = "annual",
  parsedts = TRUE
)
}
\arguments{
\item{startdt}{(char) query start date string of format YYYY-MM-DD}

\item{enddt}{(char) query end date string of format YYYY-MM-DD}

\item{varnames}{(char) comma delimited list of variable names to query for}

\item{summarytype}{(char) string for selecting a summary dataset to query
(daily, monthly, or annual)}

\item{stationlist}{(char vector) comma delimited list of station identifiers}
}
\value{
ghcn_df (tibble) table of query results
}
\description{
This function fetches a csv of climate observations selected from
a GHCN weather station and dataset and returns data as a tibble.
}
\details{
GHCN-daily data: https://doi.org/10.7289/V5D21VHZ
GHCN-monthly data: https://doi.org/10.7289/V5QV3JJ5
GHCN-annual data: https://doi.org/10.7289/JWPF-Y430

API documentation:
https://www.ncei.noaa.gov/support/access-data-service-api-user-documentation
}
