% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei.r
\name{get_spei}
\alias{get_spei}
\title{Get SPEI values from climatic water differential at given frequency and
monthly integration period.}
\usage{
get_spei(cwdiff, scale = 6, na.rm = FALSE, plot = TRUE, locname = "station")
}
\arguments{
\item{cwdiff}{(xts obj) monthly climatic water differential time series}

\item{scale}{(int) time scale, or number of months used to calculate SPEI}

\item{na.rm}{(bool) flag to mask NA values from the calculation (default=F)}

\item{plot}{(bool) Flag for making diagnostic plot of SPEI}

\item{locname}{(string) Name of the site, which appears in the plot}
}
\value{
spei_int (xts obj) spei time series at given integration period

spei_int_interp (xts obj) spei_int with NA values interpolated over
}
\description{
Optionally make a plot of the data
}
\details{
Default frequency is monthly (freq=12) but there may be ways to make
weekly frequency work (freq=52) if weekly PET is available.
}
